requirements.txt
factory-boy==2.12.0
pytest==4.6.3
pytest-cov==2.7.1
